backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
DAWG-Python==0.7.2
defusedxml==0.7.1
Django==4.1.7
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
drf-yasg==1.21.5
editdistance==0.6.2
fonetika==1.4.9
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Levenshtein==0.20.9
Markdown==3.4.1
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dotenv==1.0.0
python-Levenshtein==0.20.9
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
rapidfuzz==2.13.7
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
thefuzz==0.19.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
gunicorn==20.0.4
